Implemented a form that nearly completely builds Recipe objects
Currently implemented in the form:
- A generic rotating Carousel form class that can be extended
- A generic field type for forms
- An implemented field for text lines
- An implemented field for text paragraphs
- An implemented field for multiple text lines

Currentyl missing in the implementation:
- A field that allows a user to select ingredient objects

The form emits a signal containing a QVariantMap of all the passed data.
Fields are responsible for validating themselves.
Fields are not fully styled at this point.

Display boxes for Recipes now have an implemented clickable signal.

Minor changes to docs