Added abstract cache implementation for in memory cache.
Added the MemoryCache class, as well as a working implementation of the cache, database, data provider and factories in the main method.

Added a wrapper class for factories called FactoryManager.

Updated CMakeLists file for the new files.

Added UML files for db and cache.

Modified models to send their specialized signals as well as a generic one.

Moved the objecttypes enum definition file to the root of the project.