Implemented basic UX and some UI elements
Added the basic slots/signals that MainWindow will use to communicate with the data layer of the application.

Implemented an abstract view and abstract form class to serve as interfaces for all views and forms.

Implemented concrete view and form for profiles.

What's next to be implemented is an abstract/templated list model as well as a templated delegate.