Finished implementing profile view
Fully implemented profile view, as well as the profile form into the main window of the application.

Implemented an abstract view from which all views will inherit.

Implemented the listmodel and delegates necessary to display profile objects loaded from the database.

Barebones implementation that I intend to further develop and modernize the UI.