Implementation of the first core classes
Added the various interfaces that define the core classes functionalities.

Storable : The main interface that must be implemented to be an object that can be saved by the database.

AbstractStorable : Abstract class that defines the data fields required to be a storable object

AbstractIngredient : Interface that defines the functionality required of an Ingredient

Ingredient : Core class that encompasses the most important data of the software

AbstractIngredientDecorator : Interface that describes how a decorator to the Ingredient class should act

Recipe : Core class that extends the AbstractIngredientDecorator class, wraps the ingredient class and provides functionality.